refactor: . Move src/Query/*Group* to src/Query/Group/
