+++ src/IQuery.ts
@@ -1,7 +1,7 @@
 import type { TaskLayoutOptions } from './Layout/TaskLayoutOptions';
 import type { QueryLayoutOptions } from './QueryLayoutOptions';
 import type { Task } from './Task';
-import type { Grouper } from './Query/Grouper';
+import type { Grouper } from './Query/Group/Grouper';
 import type { QueryResult } from './Query/QueryResult';
 
 /**
