+++ src/Query/Filter/BacklinkField.ts
@@ -1,5 +1,5 @@
 import type { Task } from '../../Task';
-import type { GrouperFunction } from '../Grouper';
+import type { GrouperFunction } from '../Group/Grouper';
 import { FilterOrErrorMessage } from './FilterOrErrorMessage';
 import { TextField } from './TextField';
