+++ src/Query/Filter/FilenameField.ts
@@ -1,5 +1,5 @@
 import type { Task } from '../../Task';
-import type { GrouperFunction } from '../Grouper';
+import type { GrouperFunction } from '../Group/Grouper';
 import { TextField } from './TextField';
 
 /** Support the 'filename' search instruction.
