+++ src/Query/Filter/FunctionField.ts
@@ -1,6 +1,6 @@
 import type { Task } from '../../Task';
-import type { GrouperFunction } from '../Grouper';
-import { Grouper } from '../Grouper';
+import type { GrouperFunction } from '../Group/Grouper';
+import { Grouper } from '../Group/Grouper';
 import { Explanation } from '../Explain/Explanation';
 import { TaskExpression, parseAndEvaluateExpression } from '../../Scripting/TaskExpression';
 import type { QueryContext } from '../../Scripting/QueryContext';
