+++ src/Query/Filter/MultiTextField.ts
@@ -1,6 +1,6 @@
 import type { Task } from '../../Task';
 import type { IStringMatcher } from '../Matchers/IStringMatcher';
-import { Grouper } from '../Grouper';
+import { Grouper } from '../Group/Grouper';
 import { TextField } from './TextField';
 import type { FilterFunction } from './Filter';
