+++ src/Query/Filter/RecurringField.ts
@@ -1,5 +1,5 @@
 import type { Task } from '../../Task';
-import type { GrouperFunction } from '../Grouper';
+import type { GrouperFunction } from '../Group/Grouper';
 import type { Comparator } from '../Sorter';
 import { FilterInstructionsBasedField } from './FilterInstructionsBasedField';
