+++ src/Query/Filter/RootField.ts
@@ -1,5 +1,5 @@
 import type { Task } from '../../Task';
-import type { GrouperFunction } from '../Grouper';
+import type { GrouperFunction } from '../Group/Grouper';
 import { TextField } from './TextField';
 
 export class RootField extends TextField {
