+++ src/Query/Filter/StatusField.ts
@@ -1,6 +1,6 @@
 import type { Task } from '../../Task';
 import type { Comparator } from '../Sorter';
-import type { GrouperFunction } from '../Grouper';
+import type { GrouperFunction } from '../Group/Grouper';
 import { FilterInstructionsBasedField } from './FilterInstructionsBasedField';
 
 export class StatusField extends FilterInstructionsBasedField {
