+++ src/Query/Filter/StatusTypeField.ts
@@ -1,5 +1,5 @@
 import type { Task } from '../../Task';
-import type { GrouperFunction } from '../Grouper';
+import type { GrouperFunction } from '../Group/Grouper';
 import { StatusType } from '../../Statuses/StatusConfiguration';
 import type { Comparator } from '../Sorter';
 import { Explanation } from '../Explain/Explanation';
