+++ src/Query/Group/GroupDisplayHeadingSelector.ts
