+++ src/Query/Group/TaskGroupingTree.ts
@@ -1,8 +1,8 @@
-import type { Task } from '../Task';
+import type { Task } from '../../Task';
+import type { SearchInfo } from '../SearchInfo';
 import type { Grouper } from './Grouper';
 import { GroupingTreeNode } from './GroupingTreeNode';
 import type { TaskGroupingTreeStorage } from './TaskGroupingTreeStorage';
-import type { SearchInfo } from './SearchInfo';
 
 /*
  * A tree of tasks where every level in the tree corresponds to a grouping property.
