+++ src/Query/Group/TaskGroupingTreeStorage.ts
@@ -1,4 +1,4 @@
-import type { Task } from '../Task';
+import type { Task } from '../../Task';
 
 /**
  * Storage used for the initial grouping together of tasks.
