+++ src/Query/Group/TaskGroups.ts
@@ -1,10 +1,10 @@
-import type { Task } from '../Task';
+import type { Task } from '../../Task';
+import type { SearchInfo } from '../SearchInfo';
 import type { Grouper } from './Grouper';
 import { GroupDisplayHeadingSelector } from './GroupDisplayHeadingSelector';
 import { TaskGroupingTree } from './TaskGroupingTree';
 import { TaskGroup } from './TaskGroup';
 import type { TaskGroupingTreeStorage } from './TaskGroupingTreeStorage';
-import type { SearchInfo } from './SearchInfo';
 
 /**
  * Store all the groups of tasks generated by any 'group by'
