+++ src/Query/Query.ts
@@ -9,9 +9,9 @@ import { errorMessageForException } from '../lib/ExceptionTools';
 import { logging } from '../lib/logging';
 import { Sort } from './Sort';
 import type { Sorter } from './Sorter';
-import { TaskGroups } from './TaskGroups';
+import { TaskGroups } from './Group/TaskGroups';
 import * as FilterParser from './FilterParser';
-import type { Grouper } from './Grouper';
+import type { Grouper } from './Group/Grouper';
 import type { Filter } from './Filter/Filter';
 import { QueryResult } from './QueryResult';
 import { scan } from './Scanner';
