+++ tests/Query/Filter/TagsField.test.ts
@@ -6,8 +6,8 @@ import { DescriptionField } from '../../../src/Query/Filter/DescriptionField';
 import { fromLine } from '../../TestHelpers';
 import { TaskBuilder } from '../../TestingTools/TaskBuilder';
 import { expectTaskComparesAfter, expectTaskComparesBefore } from '../../CustomMatchers/CustomMatchersForSorting';
-import type { Grouper } from '../../../src/Query/Grouper';
-import { TaskGroups } from '../../../src/Query/TaskGroups';
+import type { Grouper } from '../../../src/Query/Group/Grouper';
+import { TaskGroups } from '../../../src/Query/Group/TaskGroups';
 import { SearchInfo } from '../../../src/Query/SearchInfo';
 import { sortBy } from '../../TestingTools/SortingTestHelpers';
